updated installer function, now supporting options when installing plugins

+++ install_plugins.sh
@@ -1,13 +1,14 @@
-version=2.3.1
+version=3.2
 
 #----------------------------------------------------------------------------------------------------------
 #
 # usage: install_plugin <plugin_uri> [<plugin_name>] [<branch_name>]
 #
 # this will install plugin from the given uri. If the optional plugin_name is given it will be used.
-# Wgen the optional branch name is given the plugin will be git checked out to that branch
+# When the optional branch name is given the plugin will be git checked out to that branch
 #
 install_plugin () {
+
 	if [ ! $1 ]
 		then
 		echo 'no plugin repository given to install - aborting!'
@@ -20,34 +21,46 @@ install_plugin () {
 		plugin_name=$2
 	fi
 
-	if [ -d $plugin_name ]
+	if [ $force_reinstall ]
 		then
-		echo "==> Plugin $plugin_name is re-installed."
-		rm -rf $plugin_name
+		if [ -d $plugin_name ]
+			then
+			echo "==> re-installing $plugin_name"
+			[ $testing ] || rm -rf $plugin_name
+		fi
 	fi
 
-	if [ $3 ]
+	if [ ! -d $plugin_name ] || [ $force_reinstall ]
 		then
-		echo "==> GIT checking out $3"
-		git clone $plugin $plugin_name -b $3
-#		git submodule add $plugin $plugin_name
-	else
-		git clone $plugin $plugin_name
-#		git submodule add $plugin $plugin_name
-	fi
+		if [ $3 ] #there is a specific branch/tag given
+			then
+			if [ $submodules ] # if the submodule switch is present install plugin as submodule
+				then
+				echo "==> add submodule $plugin as $plugin_name with branch $3"
+				[ $testing ] || git submodule add --force -b $3 $plugin $plugin_name
+			else
+				echo "==> clone repository $plugin as $plugin_name with branch $3"
+				[ $testing ] || git clone $plugin $plugin_name -b $3
+				[ $testing ] || rm -rf $plugin_name/.git # removing plugins GIT heritage to make it part of the qmplus repository
+			fi
+		else
+			if [ $submodules ]
+				then
+				echo "==> add submodule $plugin as $plugin_name"
+				[ $testing ] || git submodule add --force $plugin $plugin_name
+			else
+				echo "==> clone repository $plugin as $plugin_name"
+				[ $testing ] || git clone $plugin $plugin_name
+				[ $testing ] || rm -rf $plugin_name/.git #
+			fi
 
-#	Removing .git directory for each plugin thereby creating one big repo
-	if [ -d $plugin_name ]
-		then
-		cd $plugin_name
-		rm -rf .git
-		cd ..
+		fi
 	fi
-
 }
 
 #---------------------------------------------------------------------------------------------------------
-chdir() {
+chdir () {
+	cd $baseurl
 	if [  $1 ]
 		then
 		cd $1
@@ -58,6 +71,24 @@ chdir() {
 }
 
 #=========================================================================================================
+
+# get options
+while getopts :fst x; do
+    case $x in
+        f)
+            force_reinstall=1
+            ;;
+        s)
+            submodules=1
+            ;;
+        t)
+            testing=1
+            ;;
+    esac
+done
+shift $((OPTIND-1))
+
+
 baseurl=$(pwd)
 echo " "
 echo "install all qmplus plugins into baseurl = $baseurl (v.$version)"
@@ -65,34 +96,20 @@ echo "--------------------------------------------------------------------------
 start_time=`date +%s`
 
 #------------------------------------------------------------------------------------------
-cd $baseurl
-cd admin/tool/
-echo " "
-echo "--> " $(pwd)
-echo "-----------------------------------------------------------------"
+chdir admin/tool/
 install_plugin git@github.com:ndunand/moodle-tool_mergeusers.git mergeusers
 
 #------------------------------------------------------------------------------------------
-cd $baseurl
-cd auth/
-echo " "
-echo "--> " $(pwd)
-echo "-----------------------------------------------------------------"
+chdir auth/
 install_plugin git@github.com:ULCC-QMUL/moodle-auth_dbsyncother.git dbsyncother
 install_plugin git@github.com:catalyst/moodle-auth_saml2.git saml2
-
 install_plugin git@github.com:Microsoft/moodle-auth_oidc.git oidc
 install_plugin git@github.com:QMUL/moodle-auth_cas_db.git cas_db
 install_plugin git@github.com:QMUL/moodle-auth_ulcc_sharedsecret_sso.git ulcc_sharedsecret_sso
 
 #------------------------------------------------------------------------------------------
-cd $baseurl
-cd blocks/
-echo " "
-echo "--> " $(pwd)
-echo "-----------------------------------------------------------------"
+chdir blocks/
 install_plugin git@github.com:ULCC-QMUL/moodle-block_landingpage.git landingpage
-
 install_plugin git@github.com:ULCC-QMUL/talis-block_aspirelists.git aspirelists
 install_plugin git@github.com:ULCC-QMUL/moodle-block_assessment_information.git assessment_information master
 install_plugin git@github.com:danmarsden/moodle-block_attendance.git attendance
@@ -125,7 +142,6 @@ install_plugin git@github.com:QMUL/moodle-block_ulcc_diagnostics.git ulcc_diagno
 install_plugin git@github.com:ULCC-QMUL/moodle-block_widgets.git widgets
 install_plugin git@github.com:moodlehq/moodle-block_messages.git messages
 install_plugin git@github.com:QMUL/moodle-block_accessibility.git accessibility
-install_plugin git@github.com:QMUL/moodle-block_course_menu.git course_menu
 install_plugin git@github.com:moodlehq/moodle-block_course_overview.git course_overview
 install_plugin git@github.com:QMUL/moodle-block_course_reports.git course_reports
 install_plugin git@github.com:deraadt/moodle-block_heatmap.git heatmap
@@ -139,33 +155,29 @@ install_plugin git@github.com:QMUL/moodle-block_checklist.git checklist
 install_plugin git@github.com:QMUL/moodle-block_course_contents.git course_contents
 
 #------------------------------------------------------------------------------------------
-cd $baseurl
-cd course/format/
-echo " "
-echo "--> " $(pwd)
-echo "-----------------------------------------------------------------"
+chdir course/format/
 install_plugin git@github.com:ULCC-QMUL/moodle-course_format_qmultopics.git qmultopics develop_32
 install_plugin git@github.com:ULCC-QMUL/moodle-format_landingpage.git landingpage
 #install_plugin git@github.com:gjb2048/moodle-format_topcoll.git topcoll master
-install_plugin git@github.com:QMUL/moodle-format_topcoll.git topcoll
+#install_plugin git@github.com:QMUL/moodle-format_topcoll.git topcoll
+install_plugin git@github.com:gjb2048/moodle-format_topcoll.git topcoll
+install_plugin git@github.com:marinaglancy/moodle-format_flexsections.git flexsections
+#install_plugin git@github.com:ULCC-QMUL/moodle-format_grid.git grid develop_32
+install_plugin git@github.com:QMUL/moodle-format_grid.git grid
+install_plugin git@github.com:ULCC-QMUL/moodle-format_qmulgrid.git qmulgrid develop_32
+#install_plugin git@github.com:ULCC-QMUL/moodle-course_format_qmultc.git qmultc develop_32
+install_plugin git@github.com:ULCC-QMUL/moodle-course_format_qmultc.git qmultc qmplus34
+install_plugin git@github.com:ULCC-QMUL/moodle-course_format_qmulweeks.git qmulweeks develop_32
 
 #------------------------------------------------------------------------------------------
-cd $baseurl
-cd enrol/
-echo " "
-echo "--> " $(pwd)
-echo "-----------------------------------------------------------------"
+chdir enrol/
 #install_plugin git@github.com:bynare/moodle-enrol_auto.git auto
 install_plugin git@github.com:QMUL/moodle-enrol_auto.git auto
 install_plugin git@github.com:ULCC-QMUL/moodle-enrol_cohortcateg.git cohortcateg
 install_plugin git@github.com:QMUL/moodle-enrol_databaseextended.git databaseextended
 
 #------------------------------------------------------------------------------------------
-cd $baseurl
-cd filter/
-echo " "
-echo "--> " $(pwd)
-echo "-----------------------------------------------------------------"
+chdir filter/
 install_plugin git@github.com:ULCC-QMUL/moodle-filter_easychem.git easychem
 install_plugin git@github.com:QMUL/moode-filer_oembed.git oembed
 install_plugin git@github.com:geoffrowland/moodle-filter_jmol.git jmol
@@ -174,55 +186,31 @@ install_plugin git@github.com:QMUL/moodle-filter_kaltura.git kaltura
 install_plugin git@github.com:QMUL/moodle-filter_replace.git replace
 
 #------------------------------------------------------------------------------------------
-cd $baseurl
-cd grade/export/
-echo " "
-echo "--> " $(pwd)
-echo "-----------------------------------------------------------------"
+chdir grade/export/
 install_plugin git@github.com:davosmith/moodle-grade_checklist.git checklist
 
 #------------------------------------------------------------------------------------------
-cd $baseurl
-cd grade/grading/form
-echo " "
-echo "--> " $(pwd)
-echo "-----------------------------------------------------------------"
+chdir grade/grading/form
 #install_plugin git@github.com:marcusgreen/moodle-gradingform_btec.git btec
 install_plugin git@github.com:QMUL/moodle-gradingform_btec.git btec
 
 #------------------------------------------------------------------------------------------
-cd $baseurl
-cd grade/import/
-echo " "
-echo "--> " $(pwd)
-echo "-----------------------------------------------------------------"
+chdir grade/import/
 install_plugin git@github.com:ULCC-QMUL/moodle-gradeimport_directpearsonjson.git directpearsonjson
 
 #------------------------------------------------------------------------------------------
-cd $baseurl
-cd grade/report
-echo " "
-echo "--> " $(pwd)
-echo "-----------------------------------------------------------------"
+chdir grade/report
 install_plugin git@github.com:ULCC-QMUL/moodle-gradereport_marking.git marking
 install_plugin git@github.com:ULCC-QMUL/moodle-gradereport_qmul_sits.git qmul_sits
 
 #------------------------------------------------------------------------------------------
-cd $baseurl
-cd lib/editor/tinymce/plugins
-echo " "
-echo "--> " $(pwd)
-echo "-----------------------------------------------------------------"
+chdir lib/editor/tinymce/plugins
 #install_plugin git@github.com:kaltura/moodle-tinymce_kalturamedia.git kalturamedia
 install_plugin git@github.com:QMUL/moodle-editor_tinymce_kalturamedia.git kalturamedia
 install_plugin git@github.com:ULCC-QMUL/moodle-editor_tinymce_mathslate.git mathslate
 
 #------------------------------------------------------------------------------------------
-cd $baseurl
-cd local/
-echo " "
-echo "--> " $(pwd)
-echo "-----------------------------------------------------------------"
+chdir local/
 install_plugin git@github.com:ULCC-QMUL/moodle-local_landingpages.git landingpages
 #install_plugin git@github.com:ULCC-QMUL/moodle-local_qmframework.git qmframework
 install_plugin git@github.com:QMUL/moodle-local_qmframework.git qmframework
@@ -255,11 +243,7 @@ install_plugin git@github.com:ULCC-QMUL/moodle-local_widgets.git widgets
 install_plugin git@github.com:QMUL/moodle-local_mobile.git mobile
 
 #------------------------------------------------------------------------------------------
-cd $baseurl
-cd mod/
-echo " "
-echo "--> " $(pwd)
-echo "-----------------------------------------------------------------"
+chdir mod/
 install_plugin git@github.com:danmarsden/moodle-mod_attendance.git attendance
 install_plugin git@github.com:ULCC-QMUL/moodle-mod_turningtech.git turningtech
 install_plugin git@github.com:QMUL/moodle-mod_checklist.git checklist
@@ -296,38 +280,22 @@ install_plugin git@github.com:turnitin/moodle-mod_turnitintooltwo.git turnitinto
 install_plugin git@github.com:QMUL/moodle-mod_lightboxgallery.git lightboxgallery
 
 #------------------------------------------------------------------------------------------
-cd $baseurl
-cd mod/assign/feedback
-echo " "
-echo "--> " $(pwd)
-echo "-----------------------------------------------------------------"
+chdir mod/assign/feedback
 install_plugin git@github.com:Microsoft/moodle-assignfeedback_onenote.git onenote
 
 #------------------------------------------------------------------------------------------
-cd $baseurl
-cd mod/assign/submission
-echo " "
-echo "--> " $(pwd)
-echo "-----------------------------------------------------------------"
+chdir mod/assign/submission
 install_plugin git@github.com:pauln/moodle-assignsubmission_onlineaudio.git onlineaudio
 install_plugin git@github.com:ULCC-QMUL/moodle-mod-assign-submission_qmcw_coversheet.git qmcw_coversheet
 install_plugin git@github.com:Microsoft/moodle-assignsubmission_onenote.git onenote
 install_plugin git@github.com:QMUL/moodle-assignsubmission_mahara.git mahara
 
 #------------------------------------------------------------------------------------------
-cd $baseurl
-cd mod/quiz/accessrule/
-echo " "
-echo "--> " $(pwd)
-echo "-----------------------------------------------------------------"
+chdir mod/quiz/accessrule/
 install_plugin git@github.com:moodleou/moodle-quizaccess_honestycheck.git honestycheck
 
 #------------------------------------------------------------------------------------------
-cd $baseurl
-cd question/behaviour/
-echo " "
-echo "--> " $(pwd)
-echo "-----------------------------------------------------------------"
+chdir question/behaviour/
 install_plugin git@github.com:QMUL/moodle-qbehaviour_adaptivehints.git adaptivehints
 install_plugin git@github.com:QMUL/moodle-qbehaviour_adaptivehintsnopenalties.git adaptivehintsnopenalties
 install_plugin git@github.com:maths/moodle-qbehaviour_adaptivemultipart.git adaptivemultipart
@@ -337,19 +305,11 @@ install_plugin git@github.com:QMUL/moodle-qbehaviour_interactivehints.git intera
 install_plugin git@github.com:moodleou/moodle-qbehaviour_opaque.git opaque
 
 #------------------------------------------------------------------------------------------
-cd $baseurl
-cd question/format/
-echo " "
-echo "--> " $(pwd)
-echo "-----------------------------------------------------------------"
+chdir question/format/
 install_plugin git@github.com:ecampbell/moodle-qformat_wordtable.git wordtable
 
 #------------------------------------------------------------------------------------------
-cd $baseurl
-cd question/type/
-echo " "
-echo "--> " $(pwd)
-echo "-----------------------------------------------------------------"
+chdir question/type/
 install_plugin git@github.com:jmvedrine/moodle-qtype_jme.git jme
 install_plugin git@github.com:jmvedrine/moodle-qtype_multichoiceset.git multichoiceset
 install_plugin git@github.com:ndunand/moodle-qtype_multinumerical.git multinumerical
@@ -365,20 +325,12 @@ install_plugin git@github.com:QMUL/moodle-qtype_gapfill.git gapfill
 install_plugin git@github.com:QMUL/moodle-qtype_ddmatch.git ddmatch
 
 #------------------------------------------------------------------------------------------
-cd $baseurl
-cd question/type/
-echo " "
-echo "--> " $(pwd)
-echo "-----------------------------------------------------------------"
+chdir question/type/
 install_plugin git@github.com:QMUL/moodle-qtype_order.git order
 install_plugin git@github.com:moodleou/moodle-qtype_pmatchjme.git pmatchjme
 
 #------------------------------------------------------------------------------------------
-cd $baseurl
-cd lib/editor/atto/plugins
-echo " "
-echo "--> " $(pwd)
-echo "-----------------------------------------------------------------"
+chdir lib/editor/atto/plugins
 install_plugin git@github.com:ULCC-QMUL/moodle-editor_atto_qmulcolumns.git qmulcolumns
 install_plugin git@github.com:ULCC-QMUL/moodle-editor_atto_qmulcite.git qmulcite
 install_plugin git@github.com:ULCC-QMUL/moodle-editor_atto_qmulicons.git qmulicons
@@ -398,19 +350,11 @@ install_plugin git@github.com:ndunand/moodle-atto_morefontcolors.git morefontcol
 install_plugin git@github.com:ULCC-QMUL/kal_lib_editor_atto_plugins_kalturamedia.git kalturamedia
 
 #------------------------------------------------------------------------------------------
-cd $baseurl
-cd plagiarism/
-echo " "
-echo "--> " $(pwd)
-echo "-----------------------------------------------------------------"
+chdir plagiarism/
 install_plugin git@github.com:QMUL/moodle-plagiarism_turnitin.git turnitin
 
 #------------------------------------------------------------------------------------------
-cd $baseurl
-cd report/
-echo " "
-echo "--> " $(pwd)
-echo "-----------------------------------------------------------------"
+chdir report/
 install_plugin git@github.com:ULCC-QMUL/moodle-report_qmplus.git qmplus
 install_plugin git@github.com:pauln/moodle-report_componentgrades.git componentgrades
 #install_plugin git@github.com:ctchanandy/moodle-report_forumgraph.git forumgraph
@@ -419,26 +363,19 @@ install_plugin git@github.com:QMUL/moodle-report_forumgraph.git forumgraph
 install_plugin git@github.com:QMUL/moodle-report_overviewstatistics.git overviewstats
 
 #------------------------------------------------------------------------------------------
-cd $baseurl
-cd repository/
-echo " "
-echo "--> " $(pwd)
-echo "-----------------------------------------------------------------"
+chdir repository/
 install_plugin git@github.com:Microsoft/moodle-repository_office365.git office365
 #install_plugin git@github.com:Microsoft/moodle-repository_onenote.git onenote
 install_plugin git@github.com:QMUL/moodle-repository_evernote.git evernote
 
 #------------------------------------------------------------------------------------------
-cd $baseurl
-cd theme/
-echo " "
-echo "--> " $(pwd)
-echo "-----------------------------------------------------------------"
+chdir theme/
 #install_plugin git@github.com:apnet/moodle-theme_bcu.git bcu
 #install_plugin git@github.com:QMUL/moodle-theme_bloom.git bloom
 install_plugin git@github.com:gjb2048/moodle-theme_essential.git essential
 install_plugin git@github.com:ULCC-QMUL/moodle-theme_synergy_bootstrap.git synergy_bootstrap
-install_plugin git@github.com:ULCC-QMUL/moodle-theme_qmul.git qmul develop_32
+#install_plugin git@github.com:ULCC-QMUL/moodle-theme_qmul.git qmul develop_32
+install_plugin git@github.com:ULCC-QMUL/moodle-theme_qmul.git qmul qmplus34
 install_plugin git@github.com:ULCC-QMUL/moodle-theme_qmul_humanities.git qmul_humanities develop_32
 install_plugin git@github.com:ULCC-QMUL/moodle-theme_qmul_learning.git qmul_learning develop_32
 install_plugin git@github.com:ULCC-QMUL/moodle-theme_qmul_lifesciences.git qmul_lifesciences develop_32
@@ -448,16 +385,19 @@ install_plugin git@github.com:QMUL/moodle-theme_adaptable.git adaptable
 install_plugin git@github.com:QMUL/moodle-theme_klass.git klass
 
 #------------------------------------------------------------------------------------------
-cd $baseurl
-cd user/profile/field/
-echo " "
-echo "--> " $(pwd)
-echo "-----------------------------------------------------------------"
+chdir user/profile/field/
 install_plugin git@github.com:QMUL/moodle-profilefield_o365.git o365
 install_plugin git@github.com:QMUL/moodle-profilefield_oidc.git oidc
 
 
 cd $baseurl
+
+# update submodules if present
+if [ $submodules ]
+	then
+	echo "==> Updating submodules..."
+	[ $testing ] || git submodule update --init --recursive
+fi
 echo " "
 echo "====> DONE after $((`date +%s` - start_time)) seconds!"
 echo " "
